Add 1.3.4 -> 1.3.6 migration script
